multi: add separator widget
- add separator widget and replace repetitive line implementation on pages
- move code that will be removed when all pages have been updated to the mockup designs to a deprecated file.
- move general page components such as the side nav and top bar to a
  component.go file.
