+++ ui/decredmaterial/editor.go
@@ -69,7 +69,7 @@ func (t *Theme) Editor(editor *widget.Editor, hint string) Editor {
 		TitleLabel:        t.Body2(""),
 		IsTitleLabel:      true,
 		Bordered:          true,
-		LineColor:         t.Color.BorderColor,
+		LineColor:         t.Color.Gray1,
 		TitleLabelColor:   t.Color.IconColor,
 		errorLabel:        errorLabel,
 		requiredErrorText: "Field is required",
@@ -235,13 +235,13 @@ func (e Editor) handleEvents() {
 	if e.errorLabel.Text != "" {
 		e.LineColor = e.t.Color.Danger
 	} else {
-		e.LineColor = e.t.Color.BorderColor
+		e.LineColor = e.t.Color.Gray1
 	}
 
 	if e.requiredErrorText != "" {
 		e.LineColor = e.t.Color.Danger
 	} else {
-		e.LineColor = e.t.Color.BorderColor
+		e.LineColor = e.t.Color.Gray1
 	}
 }
