+++ ui/decredmaterial/theme.go
@@ -56,10 +56,9 @@ type Theme struct {
 		DeepBlue     color.NRGBA
 		LightGray    color.NRGBA
 		IconColor    color.NRGBA
-		BorderColor  color.NRGBA
 		InactiveGray color.NRGBA
-		Gray1     color.NRGBA
-		Gray2     color.NRGBA
+		Gray1        color.NRGBA
+		Gray2        color.NRGBA
 	}
 	Icon struct {
 		ContentCreate *widget.Icon
@@ -77,7 +76,6 @@ type Theme struct {
 	navMoreIcon           *widget.Icon
 	expandIcon            *widget.Image
 	collapseIcon          *widget.Image
-	decredIcons           map[string]image.Image
 
 	Clipboard     chan string
 	ReadClipboard chan interface{}
@@ -103,9 +101,8 @@ func NewTheme(fontCollection []text.FontFace, decredIcons map[string]image.Image
 	t.Color.Gray2 = rgb(0x899785)
 	t.Color.LightGray = rgb(0xf3f5f6)
 	t.Color.DeepBlue = rgb(0x091440)
-	t.Color.IconColor = rgb(0x3D5873)
-	t.Color.BorderColor = rgb(0xE6EAED)
-	t.Color.InactiveGray = rgb(0xC4CBD2)
+	t.Color.IconColor = rgb(0x3d5873)
+	t.Color.InactiveGray = rgb(0xc4cbd2)
 	t.Color.Black = rgb(0x000000)
 	t.TextSize = unit.Sp(16)
 
@@ -119,7 +116,6 @@ func NewTheme(fontCollection []text.FontFace, decredIcons map[string]image.Image
 	t.navigationCheckIcon = mustIcon(widget.NewIcon(icons.NavigationCheck))
 	t.dropDownIcon = mustIcon(widget.NewIcon(icons.NavigationArrowDropDown))
 
-	t.decredIcons = decredIcons
 	t.expandIcon = &widget.Image{Src: paint.NewImageOp(decredIcons["expand_icon"])}
 	t.collapseIcon = &widget.Image{Src: paint.NewImageOp(decredIcons["collapse_icon"])}
