+++ ui/util.go
@@ -1,0 +1,10 @@
+package ui
+
+import "gioui.org/widget"
+
+func mustIcon(ic *widget.Icon, err error) *widget.Icon {
+	if err != nil {
+		panic(err)
+	}
+	return ic
+}
